Pass the function value object to each function call
For functions, pass the function value object to its own function in the second parameter.

For methods, the second parameter is the receiver. For functions, it was previously unused.

Passing the function value object itself speeds up recursive calls, since no upvalues will be required if the function itself is available as a local.