Fix uninitialized memory error
Even though the shadow stack pointer will be reset at the very end of the function, the values on the stack remain.

This can lead to re-using values on the stack, so they should always be reset to nullptr.

A native local slot will be re-used and leaving that address on the stack can lead to later accessing uninitialized memory.